discard some get_fmove()

+++ src/Pokemon.cpp
@@ -62,7 +62,7 @@ Pokemon::~Pokemon()
 	erase_cmoves();
 }
 
-const Move *Pokemon::get_fmove(int) const
+const Move *Pokemon::get_fmove(unsigned) const
 {
 	return &fmove;
 }
@@ -75,9 +75,9 @@ void Pokemon::add_fmove(const Move *t_move)
 	}
 }
 
-const Move *Pokemon::get_cmove(int t_index) const
+const Move *Pokemon::get_cmove(unsigned t_index) const
 {
-	if (0 <= t_index && static_cast<unsigned>(t_index) < cmoves_count)
+	if (0 <= t_index && t_index < cmoves_count)
 	{
 		return cmoves + t_index;
 	}
